+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changelog
 These are the changes brought in each versions:
 
-v1.0.9 (TBD)
+v1.0.9 (2020-11-10)
 ===================
 Non-breaking changes:
 ---------------------